consolidate CSP header tests
